Move UbloxFirmware6 class to its own files.
